% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimates}
\alias{estimates}
\title{estimates}
\format{
data.frame, long
}
\usage{
data(estimates)
}
\description{
A data.frame containing the parameter estimates for the analysis models of interest.
}
\details{
The data.frame columns are:

\itemize{
\item \code{term} - the analysis model parameter identifier
\item \code{method} - the missing data treatment used
\item \code{variable} - the outcome measures to consider
\item \code{value} - the values to plot
}
}
\keyword{datasets}
