% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imp_methods}
\alias{imp_methods}
\title{imp_methods}
\format{
vector
}
\usage{
data(imp_methods)
}
\description{
A vector containing the names of the imputation methods used for every variable
}
\keyword{datasets}
