% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imps_ggplot}
\alias{imps_ggplot}
\title{imps_ggplot}
\format{
vector
}
\usage{
data(imps_ggplot)
}
\description{
A data.frame containing observed values and imputations in long form.
}
\details{
These data can be used to plot the density distributions of observed and imputed variables.
It was obtained by extracting the information needed for this type of plot from the mids objects produced by mice after imputation.
}
\keyword{datasets}
