% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mids_chains}
\alias{mids_chains}
\title{mids_chains}
\format{
data.frame
}
\usage{
data(mids_chains)
}
\description{
A data.frame containing chain means and standard deviations for plotting trace plots with ggplot
}
\keyword{datasets}
