% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mids_miexpert}
\alias{mids_miexpert}
\title{mids_miexpert}
\format{
mids
}
\usage{
data(mids_miexpert)
}
\description{
A mids object containing the information about the imputation performed with MI-expert
}
\details{
See \link[mice]{mids-class} help file for details.
}
\keyword{datasets}
