% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mids_migspcr}
\alias{mids_migspcr}
\title{mids_migspcr}
\format{
mids
}
\usage{
data(mids_migspcr)
}
\description{
A mids object containing the information about the imputation performed with MI-GSPCR
}
\details{
See \link[mice]{mids-class} help file for details.
}
\keyword{datasets}
