% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densities.R
\name{plot_densities}
\alias{plot_densities}
\title{Plot densities of the imputed values}
\usage{
plot_densities(data, var)
}
\arguments{
\item{data}{data.frame in long form (same as \link[plotmigspcr]{imps_ggplot})}

\item{var}{Character vector indicating which variable should be plotted.}
}
\value{
Returns the ggplot
}
\description{
Generate density plots with ggplot.
}
\examples{
plot_densities(
    data = imps_ggplot,
    var = "v1"
)
}
\author{
Edoardo Costantini, 2023
}
