% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot imputation chain summaries over iterations}
\usage{
plot_trace(data, var)
}
\arguments{
\item{data}{data.frame in long form (same as \link[plotmigspcr]{mids_chains})}

\item{var}{Character vector indicating which variable should be plotted.}
}
\value{
Returns the ggplot
}
\description{
Generate trace plots with ggplot.
}
\examples{
plot_trace(
    data = mids_chains,
    var = "v1"
)
}
\author{
Edoardo Costantini, 2023
}
