% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmigspcr-package.R
\docType{package}
\name{plotmigspcr-package}
\alias{plotmigspcr}
\alias{plotmigspcr-package}
\title{plotmigspcr: Plot mi-pca results}
\description{
A shiny app to interact with the results of the mi-pca study.
}
\author{
\strong{Maintainer}: Edoardo Costantini \email{e.costantini@tilburguniversity.edu} (\href{https://orcid.org/0000-0001-9581-9913}{ORCID})

}
\keyword{internal}
