% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mids_chains}
\alias{mids_chains}
\title{mids_chains}
\format{
data.frame
}
\usage{
data(mids_chains)
}
\description{
A data.frame containing chain means and standard deviations for plotting trace plots with ggplot
}
\details{
These data can be used to plot the progression of imputation means and standard deviations over the iterations.
It was obtained by extracting the information needed for this type of plot from the mids objects produced by mice after imputation.
}
\keyword{datasets}
