% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Plot parameter estimation results}
\usage{
plot_estimates(gg_shape, terms, approach, outcome)
}
\value{
Returns the ggplot
}
\description{
Generate plots for parameter estimates.
}
\examples{
plot_estimates(
     gg_shape = estimates,
     terms = levels(estimates$term)[-1],
     approach = levels(estimates$method),
     outcome = c("estimate", "t")
)
}
\author{
Edoardo Costantini, 2023
}
